% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Check parameters for DDUST model}
\usage{
check_parameters(parameters, rec_dev_type, silent = FALSE)
}
\arguments{
\item{parameters}{list of parameters}

\item{rec_dev_type}{Type of recruitment deviations: "barycentric", "redundant" or "none"}

\item{silent}{TRUE to receive more information}
}
\value{
updated list of parameters
}
\description{
Check parameters for DDUST model
}
\examples{
parameters <- check_parameters(parameters, rec_dev_type = 'redundant')
}
