% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_mcmc}
\alias{dd_mcmc}
\title{Stock assessment data}
\format{
An object of class \code{list} of length 3.
}
\source{
NA
}
\usage{
dd_mcmc
}
\description{
Stanfit object from TMBstan::tmbstan() containing MCMC samples from a TMB model object using Stan.
}
\keyword{datasets}
