% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_mle}
\alias{dd_mle}
\title{Stock assessment data}
\format{
An object of class \code{list} of length 3.
}
\source{
NA
}
\usage{
dd_mle
}
\description{
DDUST model outputs containing the data, parameters, model and MLE optimised values.
}
\keyword{datasets}
