% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dd_out}
\alias{dd_out}
\title{Stock assessment data}
\format{
An object of class \code{list} of length 3.
}
\source{
NA
}
\usage{
dd_out
}
\description{
List of dd_mle and dd_mcmc.
}
\keyword{datasets}
