% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDUST-package.R
\docType{package}
\name{DDUST-package}
\alias{DDUST}
\alias{DDUST-package}
\title{DDUST: Delay-Difference with User Specified Timestep}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A delay-difference R package that allows the user to specify the time step used for delays and incorporates seasonal variation in recruitment, spawning, and catchability. The delay difference with user specified time step (DDUST) model allows for monthly, bimonthly, trimonthly, quadmonthly, semi-annual and annual biomass dynamics.
}
\author{
\strong{Maintainer}: Montana Wickens \email{fisheriesassessment@daf.qld.gov.au} (\href{https://orcid.org/0000-0002-7316-014X}{ORCID})

Authors:
\itemize{
  \item Robyn Lovett (\href{https://orcid.org/0009-0005-3414-0065}{ORCID})
  \item Alise Fox (\href{https://orcid.org/0009-0002-5634-0963}{ORCID})
  \item Alexander Campbell (\href{https://orcid.org/0000-0001-7151-5501}{ORCID})
}

Other contributors:
\itemize{
  \item Michael O'Neill [contributor]
  \item George Leigh [contributor]
  \item Fisheries Queensland, FQ [copyright holder, funder]
}

}
\keyword{internal}
