% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDUST_output.R
\name{DDUST_output}
\alias{DDUST_output}
\title{Condense sdreport, fit, data, parameters and map into one full report}
\usage{
DDUST_output(rep, fit, data, parameters, map, model)
}
\arguments{
\item{rep}{sdreport(model)}

\item{fit}{optim(model)}

\item{data}{data used in MakeADFun}

\item{parameters}{parameters used in MakeADFun}

\item{map}{map used in MakeADFun}

\item{model}{model created with MakeADFun}
}
\value{
A dataframe
}
\description{
Condense sdreport, fit, data, parameters and map into one full report
}
