% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DD_model.R
\name{make_DD_model}
\alias{make_DD_model}
\title{Make the DDUST TMB model using the data, parameters and map supplied}
\usage{
make_DD_model(data, parameters, map, laplace = TRUE, hessian = TRUE)
}
\arguments{
\item{data}{list of data}

\item{parameters}{list of parameters}

\item{map}{list of fixed parameters. Used in TMB::MakeADFun()}

\item{laplace}{TRUE to use laplace approximation for recruitment deviation parameters}

\item{hessian}{calculate hessian: TRUE or FALSE}
}
\value{
model object
}
\description{
Make the DDUST TMB model using the data, parameters and map supplied
}
\examples{
data <- check_data(data)
parameters <- check_parameters(parameters, rec_dev_type = 'redundant')

make_DD_model(data, parameters, map)
}
