% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohnsrho.r
\name{mohnsrho}
\alias{mohnsrho}
\title{Calculate Mohn's rho}
\usage{
mohnsrho(dd_out)
}
\arguments{
\item{dd_out}{List of dd_out. Output from retrospective(). First list element must be model with full data timeseries.}
}
\value{
A value
}
\description{
From 'The retrospective problem in sequential population analysis: An investigation using cod fishery and simulated data.' Mohn, R. (1999)
}
