% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Stock assessment data}
\format{
An object of class \code{list} of length 11.
}
\source{
NA
}
\usage{
parameters
}
\description{
List of parameters.
}
\keyword{datasets}
