% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report2df.R
\name{report2df}
\alias{report2df}
\title{Convert report output from TMB::sdreport() to dataframe}
\usage{
report2df(report)
}
\arguments{
\item{report}{A list with eight elements created by TMB::sdreport()}
}
\value{
A dataframe
}
\description{
Convert report output from TMB::sdreport() to dataframe
}
