% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective.r
\name{retrospective}
\alias{retrospective}
\title{Retrospective analysis}
\usage{
retrospective(
  data,
  parameters,
  map,
  peel = 5,
  silent = FALSE,
  laplace = TRUE,
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{data}{list of data. See ?check_data}

\item{parameters}{list of parameters}

\item{map}{list of fixed parameters}

\item{peel}{number of years to peel (int)}

\item{silent}{TRUE to silence console output}

\item{laplace}{TRUE to use laplace approximation for recruitment deviation parameters}

\item{lower}{upper bounds on the variables passed to nlminb()}

\item{upper}{lower bounds on the variables passed to nlminb()}
}
\value{
List of MLE outputs. \code{dd_mle} is a list of model inputs and outputs from nlminb().
}
\description{
Retrospective analysis
}
\examples{

dd_out <- retrospective(data, parameters, map)

}
