% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DDUST.R
\name{run_DDUST}
\alias{run_DDUST}
\title{Quickly run DDUST}
\usage{
run_DDUST(
  data,
  parameters,
  map,
  silent = FALSE,
  laplace = FALSE,
  maxit = 5e+05,
  MCMC = FALSE,
  chains = 5,
  iter = 1000,
  init = "par",
  lower = NULL,
  upper = NULL,
  ...
)
}
\arguments{
\item{data}{list of data. See ?check_data}

\item{parameters}{list of parameters}

\item{map}{list of fixed parameters}

\item{silent}{TRUE to silence console output}

\item{laplace}{FALSE to not use laplace approximation for recruitment deviation parameters}

\item{maxit}{maximum number of iterations passed to nlminb()}

\item{MCMC}{TRUE to run MCMC after MLE}

\item{chains}{number of chains, passed to tmbstan()}

\item{iter}{number of iterations, passed to tmbstan()}

\item{init}{initial values for the sampler, passed to tmbstan()}

\item{lower}{upper bounds on the variables passed to nlminb()}

\item{upper}{lower bounds on the variables passed to nlminb()}

\item{...}{other parameters to be passed to tmbstan()}
}
\value{
List of MLE and MCMC outputs. \code{dd_mle} is a list of model inputs and outputs from optim(). \code{dd_mcmc} is a list of MCMC outputs.
}
\description{
\enumerate{
\item Model data is checked with check_data().
\item Model parameters is checked with check_parameters()
\item DDUST model is formulated with make_DD_model(). Laplace approximation is used by default.
\item Model is optimised with nlminb().
\item Standard deviation report is calculated with sdreport().
\item Outputs are collated with DDUST_output().
}
}
\examples{

# Quickly run DDUST
dd_out <- run_DDUST(data, parameters, map)

# Step-by-step run DDUST
\dontrun{
# Check data and parameters
data <- check_data(data, silent = TRUE)
parameters <- check_parameters(parameters, rec_dev_type = data$rec_dev_type, silent = TRUE)

# Make TMB model
model <- TMB::MakeADFun(data = c(model = "DDUST",data),
                        parameters,
                        random = "log_R_star",
                        DLL = "DDUST_TMBExports",
                        map = map,
                        hessian = TRUE,
                        checkParameterOrder=FALSE)

# Optimise
fit <- optim(model$par, model$fn, model$gr)

# TMB report
rep <- TMB::sdreport(model)

# Collate objects
dd_mle <- DDUST_output(rep, fit, data, parameters, map, model)
}
}
