% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DDUST.R
\name{simulate_DDUST}
\alias{simulate_DDUST}
\title{Simulate DDUST model with MCMC parameters to extract derived quantities}
\usage{
simulate_DDUST(dd_mle, dd_mcmc)
}
\arguments{
\item{dd_mle}{List of MLE outputs}

\item{dd_mcmc}{List of MCMC outputs}
}
\value{
List of simulated quantities
}
\description{
Simulate DDUST model with MCMC parameters to extract derived quantities
}
\examples{
\dontrun{
dd_out <- run_DDUST(data, parameters, map, MCMC = TRUE)
dd_sim <- simulate_DDUST(dd_out$dd_mle, dd_out$dd_mcmc)
}

}
